requirements.txt
django==6.0.3
djangorestframework==3.15.2
django-cors-headers==4.7.0
drf-yasg==1.21.8
djangorestframework-simplejwt==5.5.0
python-dotenv==1.0.1
Pillow==11.1.0
aiogram==3.18.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==2.1.0
uvicorn==0.34.0
